craete and inject mutable state flow for preferences in new settings activity, refactor new setings components for easy data passing
