+++ app/src/main/java/com/nutomic/syncthingandroid/SyncthingModule.java
@@ -9,6 +9,9 @@ import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
+import kotlinx.coroutines.flow.MutableStateFlow;
+import me.zhanghai.compose.preference.PreferenceFlow_androidKt;
+import me.zhanghai.compose.preference.Preferences;
 
 @Module
 public class SyncthingModule {
@@ -23,6 +26,12 @@ public class SyncthingModule {
     @Singleton
     public SharedPreferences getPreferences() {
         return PreferenceManager.getDefaultSharedPreferences(mApp);
+    }
+
+    @Provides
+    @Singleton
+    public MutableStateFlow<Preferences> preferencesMutableStateFlow(SharedPreferences sharedPreferences) {
+        return PreferenceFlow_androidKt.createPreferenceFlow(sharedPreferences);
     }
 
     @Provides
