+++ app/src/main/java/com/nutomic/syncthingandroid/settings/SettingsActivity.kt
@@ -3,12 +3,24 @@ package com.nutomic.syncthingandroid.settings
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.remember
+import com.nutomic.syncthingandroid.SyncthingApp
 import com.nutomic.syncthingandroid.activities.SyncthingActivity
 import com.nutomic.syncthingandroid.theme.ApplicationTheme
+import jakarta.inject.Inject
+import kotlinx.coroutines.flow.MutableStateFlow
+import me.zhanghai.compose.preference.Preferences
+import me.zhanghai.compose.preference.ProvidePreferenceLocals
+
 
 class SettingsActivity : SyncthingActivity() {
 
+    @Inject
+    lateinit var prefFlow: MutableStateFlow<Preferences>
+
     override fun onCreate(savedInstanceState: Bundle?) {
+        (application as SyncthingApp).component().inject(this)
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
 
@@ -16,11 +28,30 @@ class SettingsActivity : SyncthingActivity() {
         val startDestination: SettingsRoute = SettingsRoute.fromString(routeStr)
 
         setContent {
+            val backStack = rememberSettingsNavBackStack(startDestination)
+            val navigator = remember(backStack) {
+                object : Navigator<SettingsRoute> {
+                    override fun navigateTo(route: SettingsRoute) {
+                        backStack.add(route)
+                    }
+                    override fun navigateBack() {
+                        if (backStack.size == 1) {
+                            finish()
+                        } else {
+                            backStack.removeLastOrNull()
+                        }
+                    }
+                }
+            }
+
             ApplicationTheme {
-                SettingsNavDisplay(
-                    startDestination = startDestination,
-                    onFinishActivity = { finish() },
-                )
+                CompositionLocalProvider(LocalSettingsNavigator provides navigator) {
+                    ProvidePreferenceLocals(flow = prefFlow) {
+                        SettingsNavDisplay(
+                            backStack = backStack,
+                        )
+                    }
+                }
             }
         }
     }
