+++ app/src/main/java/com/nutomic/syncthingandroid/settings/SettingsNavigation.kt
@@ -6,8 +6,6 @@ import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.animation.togetherWith
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSerializable
 import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.runtime.toMutableStateList
@@ -18,7 +16,6 @@ import androidx.navigation3.runtime.serialization.NavBackStackSerializer
 import androidx.navigation3.runtime.serialization.NavKeySerializer
 import androidx.navigation3.ui.NavDisplay
 import kotlinx.serialization.Serializable
-import me.zhanghai.compose.preference.ProvidePreferenceLocals
 
 @Serializable
 sealed interface SettingsRoute : NavKey {
@@ -91,57 +88,38 @@ fun rememberSettingsNavBackStack(startDestination: SettingsRoute): NavBackStack<
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun SettingsNavDisplay(
-    startDestination: SettingsRoute = SettingsRoute.Root,
-    onFinishActivity: () -> Unit = {}
+    backStack: NavBackStack<SettingsRoute>
 ) {
-    val backStack = rememberSettingsNavBackStack(startDestination)
-    val navigator = remember(backStack, onFinishActivity) {
-        object : Navigator<SettingsRoute> {
-            override fun navigateTo(route: SettingsRoute) {
-                backStack.add(route)
-            }
-            override fun navigateBack() {
-                if (backStack.size == 1) {
-                    onFinishActivity()
-                } else {
-                    backStack.removeLastOrNull()
-                }
-            }
-        }
-    }
+    val navigator = LocalSettingsNavigator.current
 
-    CompositionLocalProvider(LocalSettingsNavigator provides navigator) {
-        ProvidePreferenceLocals {
-            NavDisplay(
-                backStack = backStack,
-                onBack = { navigator.navigateBack() },
-                entryProvider = entryProvider {
-                    settingsRootEntry()
-                    settingsRunConditionsEntry()
-                    settingsUserInterfaceEntry()
-                    settingsBehaviorEntry()
-                    settingsSyncthingOptionsEntry()
-                    settingsImportExportEntry()
-                    settingsTroubleshootingEntry()
-                    settingsExperimentalEntry()
-                    settingsAboutEntry()
-                },
-                transitionSpec = {
-                    // Slide in from right when navigating forward
-                    slideInHorizontally(initialOffsetX = { it }) togetherWith
-                            slideOutHorizontally(targetOffsetX = { -it })
-                },
-                popTransitionSpec = {
-                    // Slide in from left when navigating back
-                    slideInHorizontally(initialOffsetX = { -it }) togetherWith
-                            slideOutHorizontally(targetOffsetX = { it })
-                },
-                predictivePopTransitionSpec = {
-                    // Slide in from left when navigating back
-                    slideInHorizontally(initialOffsetX = { -it }) togetherWith
-                            slideOutHorizontally(targetOffsetX = { it })
-                },
-            )
-        }
-    }
+    NavDisplay(
+        backStack = backStack,
+        onBack = { navigator.navigateBack() },
+        entryProvider = entryProvider {
+            settingsRootEntry()
+            settingsRunConditionsEntry()
+            settingsUserInterfaceEntry()
+            settingsBehaviorEntry()
+            settingsSyncthingOptionsEntry()
+            settingsImportExportEntry()
+            settingsTroubleshootingEntry()
+            settingsExperimentalEntry()
+            settingsAboutEntry()
+        },
+        transitionSpec = {
+            // Slide in from right when navigating forward
+            slideInHorizontally(initialOffsetX = { it }) togetherWith
+                    slideOutHorizontally(targetOffsetX = { -it })
+        },
+        popTransitionSpec = {
+            // Slide in from left when navigating back
+            slideInHorizontally(initialOffsetX = { -it }) togetherWith
+                    slideOutHorizontally(targetOffsetX = { it })
+        },
+        predictivePopTransitionSpec = {
+            // Slide in from left when navigating back
+            slideInHorizontally(initialOffsetX = { -it }) togetherWith
+                    slideOutHorizontally(targetOffsetX = { it })
+        },
+    )
 }
